requirements.txt
blinker==1.9.0
build==1.2.2.post1
click==8.1.8
Flask==3.1.1
flask-cors==6.0.1
Flask-SQLAlchemy==3.1.1
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
pyproject_hooks==1.2.0
python-dotenv==1.1.1
SQLAlchemy==2.0.41
tomli==2.2.1
typing_extensions==4.14.0
waitress==3.0.2
Werkzeug==3.1.3
zipp==3.23.0
